requirements.txt
deepface
tf-keras
openai
opencv-python
matplotlib
python-dotenv
streamlit
watchdog

# make sure you use python 3.10:
# using the packages below will allow you to use tensorflow with the M1 chip, allowing GPU usage.
tensorflow-macos
tensorflow-metal